revert shipping controller change
